+++ src/mod/codecs/mod_g729/mod_g729.c
@@ -84,12 +84,12 @@ static switch_status switch_g729_encode(switch_codec *codec,
 										switch_codec *other_codec, 
 										void *decoded_data,
 
-										size_t decoded_data_len, 
-										int decoded_rate, 
+										uint32_t decoded_data_len, 
+										uint32_t decoded_rate, 
 										void *encoded_data,
 
-										size_t *encoded_data_len, 
-										int *encoded_rate, 
+										uint32_t *encoded_data_len, 
+										uint32_t *encoded_rate, 
 										unsigned int *flag) 
 {
 
@@ -101,7 +101,7 @@ static switch_status switch_g729_encode(switch_codec *codec,
 	}
 
 	if (decoded_data_len % 160 == 0) {
-		unsigned int new_len = 0;
+		uint32_t new_len = 0;
 		INT16 * ddp = decoded_data;
 		char *edp = encoded_data;
 		int x;
@@ -131,12 +131,12 @@ static switch_status switch_g729_decode(switch_codec *codec,
 										switch_codec *other_codec, 
 										void *encoded_data,
 
-										size_t encoded_data_len, 
-										int encoded_rate, 
+										uint32_t encoded_data_len, 
+										uint32_t encoded_rate, 
 										void *decoded_data,
 
-										size_t *decoded_data_len, 
-										int *decoded_rate, 
+										uint32_t *decoded_data_len, 
+										uint32_t *decoded_rate, 
 										unsigned int *flag) 
 {
 
@@ -175,7 +175,7 @@ static switch_status switch_g729_decode(switch_codec *codec,
 
 			int x;
 
-			unsigned int new_len = 0;
+			uint32_t new_len = 0;
 
 			test = (uint8_t *) encoded_data;
 			if (*test == 0 && *(test+1) == 0) {
